Getting ready for stuff, and hiding boostSpoolUpTime
